refactoring BeforeSuite CLI inits

+++ src/com/redhat/qe/katello/tasks/KatelloCliTasks.java
@@ -4,12 +4,9 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-
 import org.testng.Assert;
-
 import com.redhat.qe.katello.base.KatelloTestScript;
 import com.redhat.qe.katello.common.KatelloConstants;
-import com.redhat.qe.katello.common.KatelloInfo;
 import com.redhat.qe.tools.ExecCommands;
 import com.redhat.qe.tools.SSHCommandResult;
 import com.redhat.qe.tools.SSHCommandRunner;
@@ -52,69 +49,55 @@ public class KatelloCliTasks implements KatelloConstants{
 	
 	public void config_cli(String servername){
 		SSHCommandResult ssh_res;
-		String platform = execute_remote("python -c 'from platform import platform; print platform();'").getStdout();
-		int platform_id = -1;
-		
-		for(int i=0;i<CLIENT_PLATFORMS_ALLOWED.length;i++){
-			if(platform.contains(CLIENT_PLATFORMS_ALLOWED[i][0])){
-				platform_id = i;
-				log.info(String.format("The client is running on platform: [%s]",CLIENT_PLATFORMS_ALLOWED[i][1]));
-			}
-		}
-		if(platform_id == -1){
-			log.severe(String.format("ERROR: Unsupported platform for katello client: [%s]",platform));
-			System.exit(1);
-		}
 
-		// Disable the rhsm redhat.repo from being recreated - it fails yum to work properly.
-		execute_remote("rm -f /etc/yum.repos.d/redhat.repo");
-		disableRhsmYumPlugin();
-
+		// Exit if we want to reuse the system. For DEBUG only.
 		String reuseSystem = System.getProperty("katello.cli.reuseSystem", "false");
 		if(reuseSystem.equalsIgnoreCase("true")){
 			return;
 		}
 				
+		// Install Katello yum repo under /etc/yum.repos.d/
+		installKatelloRepo();
+
+		// Clean up yum caches, install wget (if not installed)
 		execute_remote("yum clean all"); // cleanup the caches
-		execute_remote("yum -y install wget");
+		if(execute_remote("rpm -q wget").getExitCode().intValue()!=0)
+			execute_remote("yum -y install wget");
 
-		if(CLIENT_PLATFORMS_ALLOWED[platform_id][0].contains("fedora")){ // Fedora
-			execute_remote(String.format("" +
-					"wget -O /etc/yum.repos.d/fedora-katello.repo %s",YUM_REPO_FEDORA_KATELLO));
-		}
-		if(CLIENT_PLATFORMS_ALLOWED[platform_id][0].contains("redhat")){ // RHEL
-			execute_remote(String.format("" +
-					"wget -O /etc/yum.repos.d/epel-katello.repo %s",YUM_REPO_RHEL_KATELLO));
-		}
-
-		execute_remote("yum repolist && yum -y erase katello-cli*"); // listing repos is needed, gets the repodata
-		execute_remote("rm -f /etc/katello/client*"); // all kinda katello client configs
+		execute_remote("yum repolist && yum -y erase katello-cli* katello-agent"); // listing repos is needed, gets the repodata
+		execute_remote("rm -f /etc/katello/client* /etc/gofer/plugins/katelloplugin*"); // remove config files
 		
-		ssh_res = execute_remote("yum -y install katello-cli*");
-		Assert.assertEquals(ssh_res.getExitCode(), new Integer(0), 
+		ssh_res = execute_remote("yum -y install katello-cli* katello-agent");
+		Assert.assertTrue(ssh_res.getExitCode().intValue()==0, 
 				"Check: return code is 0");
-		ssh_res = execute_remote("rpm -q katello-cli");
+		ssh_res = execute_remote("rpm -q katello-cli katello-agent");
 		Assert.assertEquals(ssh_res.getExitCode(), new Integer(0), 
-				"Check: return code of `rpm -q katello-cli`");
+				"Check: return code of `rpm -q katello-cli katello-agent`");
 		
-		execute_remote("sed -i \"s/localhost.localdomain/"+
-				System.getProperty("katello.server.hostname", "localhost")+
-				"/g\" "+KATELLO_CLI_CLIENT_CONFIG);
+		execute_remote("sed -i \"s/url\\s*=.*/"+
+				"url=tcp:\\/\\/\\$\\(host):5672"+
+				"/g\" "+KATELLO_AGENT_CONFIG);
 	}
 	
 	public void config_rhsm(String servername, String candlepin_ca_crt){
 		SSHCommandResult exec_result;
+
+		// Cleanup the previous registration
 		execute_remote("subscription-manager clean"); // cleanup previous registration craps
+		
+		// Exit if we want to reuse the system. For DEBUG only.
+		String reuseSystem = System.getProperty("katello.cli.reuseSystem", "false");
+		if(reuseSystem.equalsIgnoreCase("true")){
+			return;
+		}
+		
+		// Remove possible old version of RHSM, install new
 		execute_remote("wget -O /etc/yum.repos.d/epel-subscription-manager.repo http://repos.fedorapeople.org/repos/candlepin/subscription-manager/epel-subscription-manager.repo");
 		execute_remote("yum -y erase python-rhsm subscription-manager; rm -rf /etc/rhsm/* /etc/yum.repos.d/redhat.repo");
 		exec_result = execute_remote("yum -y install python-rhsm subscription-manager");
 		Assert.assertEquals(exec_result.getExitCode().intValue(), 0, "RHSM packages should get installed");
 
-		// some assertions - we need to assure we are running the version we need!
-		exec_result = execute_remote("subscription-manager register --help");
-		Assert.assertTrue(exec_result.getStdout().contains("--org=ORG"), "Check: subscription-manager --org");
-		Assert.assertTrue(exec_result.getStdout().contains("--environment=ENVIRONMENT"), "Check: subscription-manager --environment");
-		
+		// Configure
 		execute_remote("sed -i \"s/hostname = subscription.rhn.redhat.com/" +
 				"hostname = "+servername+"/g\" /etc/rhsm/rhsm.conf");
 		execute_remote("sed -i \"s/prefix = \\/subscription/prefix = \\/katello\\/api/g\" /etc/rhsm/rhsm.conf");
@@ -125,13 +108,6 @@ public class KatelloCliTasks implements KatelloConstants{
 		execute_remote("touch /etc/rhsm/ca/candlepin-local.pem; echo -e \""+candlepin_ca_crt+"\" > /etc/rhsm/ca/candlepin-local.pem");		
 	}
 
-	protected void disableRhsmYumPlugin(){
-		execute_remote("echo -e \"[main]\nenabled=0\" > /etc/yum/pluginconf.d/subscription-manager.conf");
-	}
-	protected void enableRhsmYumPlugin(){
-		execute_remote("echo -e \"[main]\nenabled=1\" > /etc/yum/pluginconf.d/subscription-manager.conf");
-	}
-	
 	public static String run_local(boolean showLogResults, String command){
 		String out = null; String tmp_cmdFile = "/tmp/katello-"+KatelloTestScript.getUniqueID()+".sh";
 		ExecCommands localRunner = new ExecCommands();
@@ -186,6 +162,35 @@ public class KatelloCliTasks implements KatelloConstants{
 		}
 		log.severe("ERROR: Output can not be extracted for the property: ["+property+"]");
 		return null;
+	}
+	
+	private int getClientPlatformID(){		
+		String platform = execute_remote("python -c 'from platform import platform; print platform();'").getStdout();
+		int platform_id = -1;
+		
+		for(int i=0;i<CLIENT_PLATFORMS_ALLOWED.length;i++){
+			if(platform.contains(CLIENT_PLATFORMS_ALLOWED[i][0])){
+				platform_id = i;
+				log.info(String.format("The client is running on platform: [%s]",CLIENT_PLATFORMS_ALLOWED[i][1]));
+			}
+		}
+		if(platform_id == -1){
+			log.severe(String.format("ERROR: Unsupported platform for katello client: [%s]",platform));
+			System.exit(1);
+		}
+		return platform_id;
+	}
+	
+	private void installKatelloRepo(){
+		int platform_id = getClientPlatformID(); // or exit if unsupported.
+		if(CLIENT_PLATFORMS_ALLOWED[platform_id][0].contains("fedora")){ // Fedora
+			execute_remote(String.format("" +
+					"wget -O /etc/yum.repos.d/fedora-katello.repo %s",YUM_REPO_FEDORA_KATELLO));
+		}
+		if(CLIENT_PLATFORMS_ALLOWED[platform_id][0].contains("redhat")){ // RHEL
+			execute_remote(String.format("" +
+					"wget -O /etc/yum.repos.d/epel-katello.repo %s",YUM_REPO_RHEL_KATELLO));
+		}
 	}
 	
 }
